requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   outcome
    #   trio
certifi==2021.10.8
    # via urllib3
cffi==1.15.0
    # via cryptography
click==8.0.3
    # via flask
cryptography==36.0.1
    # via
    #   pyopenssl
    #   urllib3
flask==2.0.2
    # via -r requirements.in
h11==0.12.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
ifaddr==0.1.7
    # via zeroconf
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.0.1
    # via jinja2
netifaces==0.11.0
    # via -r requirements.in
outcome==1.1.0
    # via trio
pillow==9.0.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyopenssl==21.0.0
    # via urllib3
selenium==4.1.0
    # via -r requirements.in
six==1.16.0
    # via pyopenssl
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure]==1.26.7
    # via selenium
werkzeug==2.0.2
    # via flask
wsproto==1.0.0
    # via trio-websocket
zeroconf==0.38.4
    # via -r requirements.in
